requirements.txt
# for window
pycocotools-windows
# for linux
# pycocotools
tqdm
opencv-python
imagesize
scipy